third test passing

+++ src/IPlannedStart.java
@@ -3,6 +3,6 @@ import java.time.LocalDateTime;
 public interface IPlannedStart {
    public LocalDateTime getStartTime();
    public void setStartTime(LocalDateTime startTime);
-   public int getCount();
-   public void setCount(int count);
+   public long getCount();
+   public void setCount(long count);
 }
